SFT-1089: replaced export_summary_task
