+++ ports/stm32/boards/Passport/manifest.py
@@ -245,7 +245,6 @@ freeze('$(MPY_DIR)/ports/stm32/boards/Passport/modules',
         'tasks/delete_multisig_task.py',
         'tasks/delay_task.py',
         'tasks/erase_passport_task.py',
-        'tasks/export_summary_task.py',
         'tasks/format_microsd_task.py',
         'tasks/fuelgauge_task.py',
         'tasks/generate_addresses_task.py',
