+++ ports/stm32/boards/Passport/modules/tasks/__init__.py
@@ -24,7 +24,6 @@ from .delay_task import delay_task
 from .delete_multisig_task import delete_multisig_task
 from .double_check_psbt_change_task import double_check_psbt_change_task
 from .erase_passport_task import erase_passport_task
-from .export_summary_task import export_summary_task
 from .format_microsd_task import format_microsd_task
 # from .fcc_copy_files_task import fcc_copy_files_task
 from .generate_addresses_task import generate_addresses_task
